+++ admin.php
@@ -100,10 +100,7 @@ function wimbblock_admin() {
 		}
 		echo '</form>';
 	} elseif ( strpos( $active_tab, 'blocking' ) !== false ) {
-		//wimbblock_blocking_tab( $active_tab );
-
 		echo '<h3>' . esc_html( __( 'Versions Control', 'wimb-and-block' ) ) . '</h3>';
-
 		if ( is_multisite() && ! is_main_site() && is_plugin_active_for_network( WIMBBLOCK_BASENAME ) ) {
 			echo '<p>';
 			echo wp_kses_post(
@@ -145,11 +142,11 @@ function wimbblock_admin() {
 		echo '<h3>' . esc_html( __( 'WIMB Table', 'wimb-and-block' ) ) . '</h3>';
 		wimbblock_mgt_table();
 	} elseif ( $active_tab === 'mgt' ) {
-		require_once __DIR__ . '/admin/mgt-table.php';
-		echo '<h3>' . esc_html( __( 'WIMB Table Management', 'wimb-and-block' ) ) . '</h3>';
+		require_once __DIR__ . '/admin/block-unblock.php';
+		echo '<h3>' . esc_html( __( 'Block / unblock entries', 'wimb-and-block' ) ) . '</h3>';
 		wimbblock_selection_table();
 	} elseif ( $active_tab === 'wimbstat' ) {
-		require_once __DIR__ . '/admin/mgt-statistik-table.php';
+		require_once __DIR__ . '/admin/statistics-table.php';
 	} elseif ( $active_tab === 'exclude' ) {
 		if ( is_multisite() && ! is_main_site() && is_plugin_active_for_network( WIMBBLOCK_BASENAME ) ) {
 			echo '<p>';
@@ -236,12 +233,12 @@ function wimbblock_admin_tabs() {
 			'title' => __( 'WIMB Table', 'wimb-and-block' ),
 		);
 		$tabs[] = array(
-			'tab'   => 'mgt',
-			'title' => __( 'WIMB Table Management', 'wimb-and-block' ),
-		);
-		$tabs[] = array(
 			'tab'   => 'wimbstat',
 			'title' => __( 'WIMB Table Statistics', 'wimb-and-block' ),
+		);
+		$tabs[] = array(
+			'tab'   => 'mgt',
+			'title' => __( 'Block / unblock entries', 'wimb-and-block' ),
 		);
 		$tabs[] = array(
 			'tab'   => 'exclude',