+++ admin/exclude.php
@@ -37,7 +37,7 @@ function wimbblock_exclude_form() {
 
 // Sanitize and validate input. Accepts an array, return a sanitized array.
 function wimbblock_exclude_validate( $params ) {
-	//var_dump($params);wp_die();
+	// var_dump($params);wp_die();
 	if ( ! empty( $_POST ) && check_admin_referer( 'wimbblock', 'wimbblock_nonce' ) ) {
 		if ( isset( $_POST['submit'] ) ) {
 			$newparams = array();