+++ php/always-block.php
@@ -8,7 +8,6 @@
 // Direktzugriff auf diese Datei verhindern.
 defined( 'ABSPATH' ) || die();
 
-//
 function wimbblock_always( $table_name, $agent, $blocked, $id, $robots ) {
 	$alwayses = wimbblock_get_option( 'wimbblock_always' );
 	if ( $alwayses !== false ) {